Get rid of PowerPoint presenter, convert PPT to PDF on the fly
PowerPoint has never been reliable for us, consequently, it will be
removed from the presentation system as much as possible.

When you try to open a PowerPoint file, it will be converted to PDF
and then the PDF file will be presented using the rock-solid PDF
presententation code. The temporary PDF file will be deleted afterwards.