PresentationWindow: grow up and don't allow closing
The PresentationWindow used to be a child of the main window, which is
unfortunate as it will not allow the user to control the main window in
a single-monitor scenario (which is useful for debugging). Without the
parent-child relationship, things work way more in our favor.

Also, don't allow the user to just close the presentation window. We
reserve the right to manage that ourselves.