Remove the ScreenView
It's sad to see it go, because especially the win7 version is something
I spend a lot of time on and I'm still proud of eventually getting it
to work reliably enough for semi-production use. Definitely proves
that the Abitur is not hard enough.

Unfortunately, the whole thing proved to be not very useful in practice,
and Avira is pretty much convinced that it's some kind of virus.
We'll get rid of it now, and if it is really missed, nobody says
we can't bring it back one day.