PDF Presenter: use projection image for preview, too
Previously, the preview inside the window rendered the the
current page by itself, too, so that it would look nice.
As a side effect, the preview image lagged behind the projected
image, very noticably so on a slower machine.

So we traded a beautiful operator display for synchronized
page changes. I think that's a good deal.